Cleanups and moving all generic charts components to charts directory
