+++ src/components/common/AnalyticsCharts/CourseTimeCompletionChart/CourseTimeCompletionLegend.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IPieBarLegendContent } from '../../PieBarChart/pieBarChart.interface';
+import { IPieBarLegendContent } from '../../charts/PieBarChart/pieBarChart.interface';
 
 export default function AvgCompletionTimeLegend({
   barValue,
