+++ src/components/common/charts/WMHorizontalBarChart/style.module.scss
