+++ src/components/common/charts/index.ts
@@ -1,3 +1,7 @@
 import WMLineChart from './WMLineChart';
+import PieBarChart from './PieBarChart';
+import WMProgress from './WMProgress';
+import WMHorizontalBarChart from './WMHorizontalBarChart';
+import PieBarSummary from './PieBarSummary';
 
-export { WMLineChart };
+export { WMLineChart, PieBarChart, WMProgress, WMHorizontalBarChart, PieBarSummary };
